requirements.txt
anyio==4.0.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.7
Werkzeug==3.0.1
